backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic>=2.10,<3
pydantic-settings>=2.6,<3

# Supabase (2.28+ uses realtime 2.x, compatible with websockets>=13 for openai[realtime])
supabase>=2.28.0
python-jose[cryptography]==3.3.0

# LiveKit
livekit==0.11.1
livekit-api==0.6.0

# OpenAI (>=2.22 + realtime extra for client.realtime.connect WebSocket)
openai[realtime]>=2.22.0
numpy>=1.24.0

# AWS S3 (optional - for recordings if not using Supabase Storage)
boto3==1.34.110

# HTTP
httpx==0.27.0

# Async
anyio==4.3.0
